detect and disable debugger, rather than adding a noop DB::DB
When used as -d:JSON, debugger features will be enabled.  This requires
a DB::DB sub exist, and slows down many parts of perl.  Instead, we can
detect that we have been used in this way, and disable the debugger
features.  This also allows the module to be used with perl5db like
perl -d -MDevel::JSON -e'...' since we are not installing a noop DB::DB.

The code is intentionally placed before strict and warnings are enabled,
so that the debug flags are disabled by the time they are loaded and
used.